Text can be displayed on screen
As long as it's an "a" or a "b" as none of the other tiles exist yet.
Going to work on the image -> header tool next because there is an issue
in the image text display (with it not correctly going to the next line)
and I could use the tool to debug it.